+++ packages/nimble-components/src/step/tests/step.spec.ts
@@ -1,0 +1,7 @@
+import { Step, stepTag } from '..';
+
+describe('Step', () => {
+    it('can construct an element instance', () => {
+        expect(document.createElement(stepTag)).toBeInstanceOf(Step);
+    });
+});