+++ packages/nimble-components/src/stepper/template.ts
@@ -1,0 +1,6 @@
+import { html } from '@ni/fast-element';
+import type { Stepper } from '.';
+
+export const template = html<Stepper>`
+    <template>stepper: <slot></slot></template>
+`;