+++ packages/storybook/src/nimble/stepper/types.ts
@@ -1,0 +1,6 @@
+export const ExampleStepType = {
+    simple: 'simple',
+    many: 'many',
+    wide: 'wide'
+} as const;
+export type ExampleStepType = (typeof ExampleStepType)[keyof typeof ExampleStepType];